+++ src/cli/serve.ts
@@ -13,11 +13,11 @@ import { bundleHtml } from "@/cli/html";
 import { createAssetRouter, createPageRouter, createStaticRouter, withoutExt } from "@/cli/route";
 import type { Router } from "@/cli/route";
 import { bundleScriptEsbuild } from "@/cli/script";
-import { contentType } from "@/lib/core/coreutil";
 import { Link, Script } from "@/lib/core/elements";
 import { clearStore, generateStore } from "@/lib/core/store";
 import type { Store } from "@/lib/core/store";
-import { stringifyToHtml } from "@/lib/server/serverfn";
+import { contentType } from "@/lib/core/util";
+import { stringifyToHtml } from "@/lib/server/stringify";
 import { ErrorPage } from "@/page/error";
 import { hanabi_error_css } from "@/page/hanabi-error";
 import chokidar from "chokidar";
