+++ src/core.ts
@@ -1,6 +1,7 @@
 export * from "@/lib/core/component";
-export * from "@/lib/core/coreutil";
+export * from "@/lib/core/util";
 export * from "@/lib/core/design";
+export * from "@/lib/core/dom";
 export * from "@/lib/core/elements";
 export * from "@/lib/core/properties";
 export * from "@/lib/core/store";
