+++ src/lib/ui/svgIconStore.ts
@@ -1,6 +1,6 @@
 import type { Attribute, HComponentFn } from "@/lib/core/component";
-import { hash_djb2 } from "@/lib/core/coreutil";
 import type { Store } from "@/lib/core/store";
+import { hash_djb2 } from "@/lib/core/util";
 import { hSvgIconFont } from "@/lib/ui/svgIconFont";
 import type { HSvgIconArg } from "@/lib/ui/svgIconFont";
